refactor: moving ‘bootstrap’ to ‘shell’
Seems like a better name for it.
