+++ shell/index.tsx
@@ -1,1 +1,1 @@
-import('./bootstrap');
+import('./shell');
